dev: find supplier debt report

+++ book-store-management-backend/module/supplierdebt/supplierdebtmodel/supplier_debt.go
@@ -10,8 +10,8 @@ import (
 type SupplierDebt struct {
 	Id            string               `json:"id" gorm:"column:id;" example:"debt001"`
 	SupplierId    string               `json:"supplierId" gorm:"column:supplierId;" example:"123"`
-	Quantity      float32              `json:"qty" gorm:"column:qty;" example:"-70000"`
-	QuantityLeft  float32              `json:"qtyLeft" gorm:"column:qtyLeft;" example:"-100000"`
+	Quantity      int                  `json:"qty" gorm:"column:qty;" example:"-70000"`
+	QuantityLeft  int                  `json:"qtyLeft" gorm:"column:qtyLeft;" example:"-100000"`
 	DebtType      *enum.DebtType       `json:"type" gorm:"column:type;" example:"Debt"`
 	CreatedBy     string               `json:"-" gorm:"column:createdBy;"`
 	CreatedByUser usermodel.SimpleUser `json:"createdBy" gorm:"foreignKey:CreatedBy"`
